chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "docx2txt>=0.9",
    "fastapi>=0.116.1",
    "huggingface-hub>=0.34.4",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-huggingface>=0.3.1",
    "langchain>=0.3.27",
    "pymupdf>=1.26.3",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
